pyproject.toml
[tool.poetry]
name = "module-7"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
django-stubs = "^5.0.4"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
